refactor(processor): split into light root and /processor heavy entry
Importing the api-processor root used to drag in the DDL builder ->
ddlapi/parser -> libpg-query WASM, because the entry re-exports the build
engine and the parser is reachable from any import. UI main-thread code (and
any model-only consumer) paid for the parser/WASM it never uses.

Split the public surface into two explicit entries:

- "." (light): types, constants, api-type ids and pure utilities. Parser-free,
  sideEffects:false. ~5.6 kB ESM bundle with no ddlapi/parser, pgsql-parser or
  libpg-query references.
- "./processor" (heavy): PackageVersionBuilder, build strategy, compare/build
  engine and ddlBuilder. This is where actual spec processing lives.

Mechanics:
- Move ddlBuilder + the ddl.parser re-export into apitypes/ddl/ddl.builder.ts;
  the apitypes/ddl barrel (and thus the root) is now model-only. builder.ts
  imports ddlBuilder from the heavy module.
- Production parser imports (buildFromDdl, prepareDdlExtractor, DdlNonFatalError,
  DdlExtractor) move from the ddlapi root to @netcracker/qubership-apihub-ddlapi
  /parser; model symbols (Realm, DdlErrorKind) stay on the root.
- vite.config emits two ESM entries (api-processor.es.js, api-processor-engine
  .es.js) and externalizes ddlapi, libpg-query, pgsql-parser and pgsql-deparser
  so the consuming bundler/runtime owns WASM loading. Dropped the unused UMD
  format and deleted vite-libpg-query-browser.plugin.ts (WASM is no longer
  bundled/inlined here).
- package.json: sideEffects:false, dual . + ./processor exports (ESM+CJS),
  typesVersions for the /processor subpath.
- Tests that use the engine import from ./processor; the one test that parsed
  DDL imports buildFromDdl from ddlapi/parser.

Verified: tsc + vite build green; light ESM and CJS contain no parser/WASM
references; the engine entry externalizes ddlapi/parser; full unit suite passes
(1182 tests).
