+++ src/apitypes/ddl/ddl.types.ts
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-import { DdlExtractor, DdlNonFatalError, Realm } from '@netcracker/qubership-apihub-ddlapi'
+import type { Realm } from '@netcracker/qubership-apihub-ddlapi'
+import type { DdlExtractor, DdlNonFatalError } from '@netcracker/qubership-apihub-ddlapi/parser'
 
 export interface ParsedDdlData {
   // the Realm produced by buildFromDdl for this single `.sql` file (per-file, self-sufficient — D6)
