+++ src/index.ts
@@ -14,10 +14,14 @@
  * limitations under the License.
  */
 
+// Light entry: shared types, constants and pure utilities (parser-free, no WASM).
+// The spec-processing engine — PackageVersionBuilder, the compare/build engine and
+// the DDL parser — lives in the '/processor' entry. Importing this root never pulls
+// the ddlapi parser / libpg-query WASM, so main-thread/UI code can import freely.
 export * from './apitypes'
-export * from './builder'
-export * from './builder-strategy'
-export * from './components'
+// Parser-free compare helpers (e.g. calculateTotalChangeSummary). The compare/build
+// engine itself (compare, compare.ddl) is part of the '/processor' entry.
+export * from './components/compare/compare.utils'
 export * from './consts'
 export * from './types'
 export {
