+++ src/processor.ts
@@ -14,15 +14,14 @@
  * limitations under the License.
  */
 
-// Light DDL surface: model-only (parser-free). The parser-bearing pieces
-// (`parseDdlFile`, `ddlBuilder`) live in ./ddl.builder and are exposed only via
-// the package's `/processor` entry, so importing the api-processor root never
-// pulls the ddlapi parser / libpg-query WASM.
+// Heavy entry: the spec-processing engine. This is a superset of the light root
+// plus PackageVersionBuilder, the build strategy, the compare/build engine and the
+// DDL builder. It transitively imports the ddlapi parser (pgsql-parser / libpg-query
+// WASM), so import it only where actual processing happens (publish/changelog/export
+// workers, build-task-consumer) — never on the UI main thread.
 
-export * from './ddl.changes'
-export * from './ddl.consts'
-export * from './ddl.document'
-export * from './ddl.entities'
-export * from './ddl.types'
-export * from './ddl.utils'
-export * from './ddl.validation'
+export * from './index'
+export * from './builder'
+export * from './builder-strategy'
+export * from './components'
+export * from './apitypes/ddl/ddl.builder'
