+++ test/ddl-changelog.test.ts
@@ -15,7 +15,8 @@
  */
 
 import { describe, expect, test } from '@jest/globals'
-import { buildFromDdl, Realm } from '@netcracker/qubership-apihub-ddlapi'
+import { buildFromDdl } from '@netcracker/qubership-apihub-ddlapi/parser'
+import type { Realm } from '@netcracker/qubership-apihub-ddlapi'
 import { Diff } from '@netcracker/qubership-apihub-api-diff'
 import { compareDdlDocuments } from '../src/apitypes/ddl/ddl.changes'
 import { DdlComparePairContext } from '../src/types'
